Horse data and spruce update for ice mountain and ice plaints growth
Added in horse growth data below mushroom cow to allow proper spawning in biomes.  Value is set relatively low to allow for a rarity of the creatures.

Redwoods have been added to the cold regions with a 1.5 ratio putting them slightly below the warmer zones but still high for their natural habitat.